requirements.txt
discord.py
requests
python-dotenv
newspaper3k
lxml
nltk
pymongo
feedparser
googletrans
